requirements.txt
Flask==3.1.0
Flask-SQLAlchemy==3.1.1
Flask-Login==0.6.3
psycopg2-binary==2.9.10
openai==1.77.0
python-dotenv==1.1.0
Werkzeug==3.1.3
SQLAlchemy==2.0.40
